requirements.txt
# Core dependencies
flask==2.3.3
flask-cors==4.0.0
pyjwt==2.8.0
python-dotenv==1.0.0

# Claude AI
anthropic==0.34.2

# Interactive Brokers
ib_insync==0.9.86

# Database
# sqlite3 is included with Python

# Session management (optional but recommended)
redis==5.0.1

# MCP Bridge Service (for Claude Desktop integration)
aiohttp==3.12.15
aiohttp-cors==0.8.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Development
black==23.9.1
pylint==3.0.2